split to key_value_store_manager_impl.rs
